+++ tests/test-likelihood_data.R
@@ -8,6 +8,7 @@ library(gadget3)
 generate_ld <- function (tbl, all_stocks = list(), all_fleets = list(), ...) {
     if (is.character(tbl)) tbl <- read.table(text = tbl, header = TRUE, stringsAsFactors = TRUE)
     if (is.null(tbl$number)) tbl$number <- as.numeric(seq_len(nrow(tbl)))
+    all_stocks <- lapply(all_stocks, function (x) g3_stock(x, 1))
     out <- gadget3:::g3l_likelihood_data('ut', structure(tbl, ...), all_stocks = all_stocks, all_fleets = all_fleets)
 
     # NB: A failed merge would result in repeated instances
@@ -626,16 +627,24 @@ ok_group('g3l_likelihood_data:stock', {
           6 2000    b  2000.6
           4 2001    b  2001.4
           6 2001    b  2001.6
-        ")
+        ", all_stocks = c('a', 'b'))
     ok(ut_cmp_identical(dimnames(ld$number)[['stock']], c("a", "b")), "Array has stocks a & b")
     ok(ut_cmp_identical(ld$stock_map, list(a = 1L, b = 2L)), "stock_map is 1:1 mapping")
 
+    ok(ut_cmp_error(generate_ld("
+        age year stock number
+          3 1999    a  1999.3
+          4 1999    b  1999.4
+          5 1999    kapow  1999.6
+          6 1999    zot  1999.6
+        ", all_stocks = c('a', 'b')), "kapow, zot"), "Unknown stock names in data an error")
+
     # Generate a list of stocks "stock_(imm,mat)_(f,m)"
-    stocks <- lapply(paste(
+    stock_names <- paste(
         'stock',
         rep(c('imm', 'mat'), each = 2),
         c('f', 'm'),
-        sep = "_"), function (x) g3_stock(x, 1))
+        sep = "_")
     ld <- generate_ld("
         age year stock_re number
           3 1999    _f$  1999.3
@@ -645,7 +654,7 @@ ok_group('g3l_likelihood_data:stock', {
           6 2000    ^stock_imm  2000.6
           4 2001    ^stock_imm  2001.4
           6 2001    ^stock_mat  2001.6
-        ", all_stocks = stocks)
+        ", all_stocks = stock_names)
     ok(ut_cmp_identical(
         dimnames(ld$number)[['stock_re']],
         c("_f$", "^stock_mat", "^stock_imm")), "Array names are regexes")
@@ -657,12 +666,19 @@ ok_group('g3l_likelihood_data:stock', {
             stock_mat_f = 1L,
             stock_mat_m = 2L)), "Stock map used first regexes first")
 
+    ok(ut_cmp_error(generate_ld("
+        age year stock_re number
+          3 1999    _f$  1999.3
+          3 1999    _g$  1999.3
+          3 1999    _h$  1999.3
+        "), "_g\\$, _h\\$"), "Regexes that don't match anything an error")
+
     # Generate a list of stocks "stock_(imm,mat)_f" (NB: not male)
-    stocks <- lapply(paste(
+    stock_names <- paste(
         'stock',
         rep(c('imm', 'mat'), each = 2),
         c('f', 'm'),
-        sep = "_"), function (x) g3_stock(x, 1))
+        sep = "_")
     ld <- generate_ld("
         age year stock_re number
           3 1999    _mat_f$  1999.3
@@ -672,7 +688,7 @@ ok_group('g3l_likelihood_data:stock', {
           6 2000    _imm_f$  2000.6
           4 2001    _imm_f$  2001.4
           6 2001    _mat_f$  2001.6
-        ", all_stocks = stocks)
+        ", all_stocks = stock_names)
     ok(ut_cmp_identical(
         dimnames(ld$number)[['stock_re']],
         c("_mat_f$", "_imm_f$")), "Array names are regexes")